pyproject.toml
[tool.poetry]
name = "djangoproject"
version = "0.1.0"
description = ""
authors = ["Viachelav Lisovoi <viacheslav.lisovoi@onix-systems.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.2"
django-rest-framework = "^0.1.0"
PyJWT = "^2.6.0"
drf-yasg = "^1.21.5"
celery = "^5.2.7"
redis = "^4.5.4"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
